Move code that was in wrong facade and worker
