+++ api/caasunitprovisioner/client_test.go
@@ -230,3 +230,57 @@ func (s *unitprovisionerSuite) TestApplicationConfig(c *gc.C) {
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(cfg, jc.DeepEquals, application.ConfigAttributes{"foo": "bar"})
 }
+
+func (s *unitprovisionerSuite) TestUpdateUnits(c *gc.C) {
+	var called bool
+	client := newClient(func(objType string, version int, id, request string, a, result interface{}) error {
+		called = true
+		c.Check(objType, gc.Equals, "CAASUnitProvisioner")
+		c.Check(id, gc.Equals, "")
+		c.Assert(request, gc.Equals, "UpdateApplicationsUnits")
+		c.Assert(a, jc.DeepEquals, params.UpdateApplicationUnitArgs{
+			Args: []params.UpdateApplicationUnits{
+				{
+					ApplicationTag: "application-app",
+					Units: []params.ApplicationUnitParams{
+						{Id: "uuid", Address: "address", Ports: []string{"port"},
+							Status: "active", Info: "message"},
+					},
+				},
+			},
+		})
+		c.Assert(result, gc.FitsTypeOf, &params.ErrorResults{})
+		*(result.(*params.ErrorResults)) = params.ErrorResults{
+			Results: []params.ErrorResult{{}},
+		}
+		return nil
+	})
+	err := client.UpdateUnits(params.UpdateApplicationUnits{
+		ApplicationTag: names.NewApplicationTag("app").String(),
+		Units: []params.ApplicationUnitParams{
+			{Id: "uuid", Address: "address", Ports: []string{"port"},
+				Status: "active", Info: "message"},
+		},
+	})
+	c.Check(err, jc.ErrorIsNil)
+	c.Check(called, jc.IsTrue)
+}
+
+func (s *unitprovisionerSuite) TestUpdateUnitsCount(c *gc.C) {
+	client := newClient(func(objType string, version int, id, request string, a, result interface{}) error {
+		*(result.(*params.ErrorResults)) = params.ErrorResults{
+			Results: []params.ErrorResult{
+				{Error: &params.Error{Message: "FAIL"}},
+				{Error: &params.Error{Message: "FAIL"}},
+			},
+		}
+		return nil
+	})
+	err := client.UpdateUnits(params.UpdateApplicationUnits{
+		ApplicationTag: names.NewApplicationTag("app").String(),
+		Units: []params.ApplicationUnitParams{
+			{Id: "uuid", Address: "address"},
+		},
+	})
+	c.Check(err, gc.ErrorMatches, `expected 1 result\(s\), got 2`)
+}
