+++ worker/caasunitprovisioner/broker.go
@@ -3,10 +3,16 @@
 
 package caasunitprovisioner
 
-import "github.com/juju/juju/core/application"
+import (
+	"github.com/juju/juju/caas"
+	"github.com/juju/juju/core/application"
+	"github.com/juju/juju/watcher"
+)
 
 type ContainerBroker interface {
 	EnsureUnit(appName, unitName, spec string) error
+	WatchUnits(appName string) (watcher.NotifyWatcher, error)
+	Units(appName string) ([]caas.Unit, error)
 }
 
 type ServiceBroker interface {
